Fix lunit readlink usage for Mac OSX
On Mac OSX (and probably other non-Linux *nixes) lunit's shell script
uses the nonstandard option '-e' in its readlink command.

The patch which will make it use '-f' which changes the behavior only
by not requiring that the last component in the path exist. Not much
of a problem for lunit's script in which it is  known that it really
does exist.

Signed-off-by: Michael Roth <mroth@nessie.de>